move comsol tests to test_io_comsol
