+++ tests/test_io.py
@@ -642,328 +642,3 @@ def test_default_audio_subtype(default_audio_subtype_mock):
     subtype_return = pyfar.io.default_audio_subtype(format=audio_format)
     assert subtype_return == 'bla'
     default_audio_subtype_mock.assert_called_with(audio_format)
-
-
-def test_read_comsol_file_not_found():
-    """Test error of find not found."""
-    path = 'blabla.csv'
-    with pytest.raises(FileNotFoundError):
-        io.read_comsol_header(path)
-
-
-def test_read_comsol_wrong_file_type():
-    """Test read comsol file format."""
-    filename = 'bla.bla'
-    with pytest.raises(SyntaxError):
-        io.read_comsol_header(filename)
-
-
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_warning_for_db_values(type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', 'level_only')
-    with pytest.warns(Warning, match='dB'):
-        io.read_comsol(path + type)
-
-
-@pytest.mark.parametrize("filename,expressions",  [
-    ('intensity_average', ['pabe.Ix', 'pabe.Iy', 'pabe.Iz']),
-    ('intensity_only', ['pabe.Ix', 'pabe.Iy', 'pabe.Iz']),
-    ('intensity_parametric', ['pabe.Ix', 'pabe.Iy', 'pabe.Iz']),
-    ('level_only', ['pabe.Lp_t']),
-    ('pressure_acceleration_parametric_time', ['actd.p_t', 'actd.a_inst']),
-    ('pressure_only', ['pabe.p_t']),
-    ('pressure_parametric', ['pabe.p_t']),
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_header_expressions(filename, expressions, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    actual_expressions, actual_expressions_unit, parameters, domain, \
-        domain_data = io.read_comsol_header(path + type)
-    assert len(actual_expressions) == len(expressions)
-    for i, exp in enumerate(expressions):
-        assert actual_expressions[i] == expressions[i]
-
-
-@pytest.mark.parametrize("filename,expressions_unit",  [
-    ('intensity_average', ['W/m^2', 'W/m^2', 'W/m^2']),
-    ('intensity_only', ['W/m^2', 'W/m^2', 'W/m^2']),
-    ('intensity_parametric', ['W/m^2', 'W/m^2', 'W/m^2']),
-    ('level_only', ['dB']),
-    ('pressure_acceleration_parametric_time', ['Pa', 'm/s^2']),
-    ('pressure_only', ['Pa']),
-    ('pressure_parametric', ['Pa']),
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_header_expressions_unit(filename, expressions_unit, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    actual_expressions, actual_expressions_unit, parameters, domain, \
-        domain_data = io.read_comsol_header(path + type)
-    assert len(actual_expressions_unit) == len(expressions_unit)
-    for i, exp in enumerate(expressions_unit):
-        assert actual_expressions_unit[i] == expressions_unit[i]
-
-
-@pytest.mark.parametrize("filename,parameters",  [
-    ('intensity_average',
-     {'theta': [0.0, 0.7854, 1.5708, 2.3562, 3.1416],
-      'phi': [0., 1.5708, 3.1416, 4.7124]}),
-    ('intensity_only', {}),
-    ('intensity_parametric',
-     {'theta': [0.0, 0.7854, 1.5708, 2.3562, 3.1416],
-      'phi': [0., 1.5708, 3.1416, 4.7124]}),
-    ('level_only', {}),
-    ('pressure_acceleration_parametric_time',
-     {'A0': [0.5, 1., 1.5], 'f0': [50, 100, 150, 200]}),
-    ('pressure_only', {}),
-    ('pressure_parametric',
-     {'theta': [0.0, 0.7854, 1.5708, 2.3562, 3.1416],
-      'phi': [0., 1.5708, 3.1416, 4.7124]}),
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_header_parameters(filename, parameters, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, actual_parameters, actual_domain, \
-        domain_data = io.read_comsol_header(path + type)
-    assert parameters == actual_parameters
-
-
-@pytest.mark.parametrize("filename,domain",  [
-    ('intensity_average', 'freq'),
-    ('intensity_only', 'freq'),
-    ('intensity_parametric', 'freq'),
-    ('level_only', 'freq'),
-    ('pressure_acceleration_parametric_time', 'time'),
-    ('pressure_only', 'freq'),
-    ('pressure_parametric', 'freq'),
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_header_domain(filename, domain, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, actual_domain, domain_data \
-        = io.read_comsol_header(path + type)
-    assert domain == actual_domain
-
-
-@pytest.mark.parametrize("filename,domain_data",  [
-    ("intensity_average", [100, 500]),
-    ("intensity_only", [100, 500]),
-    ("intensity_parametric", [100, 500]),
-    ("level_only", [100, 500]),
-    ("pressure_acceleration_parametric_time",
-     [0, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1]),
-    ("pressure_only", [100, 500]),
-    ("pressure_parametric", [100, 500]),
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_header_domain_data(filename, domain_data, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, actual_domain_data \
-        = io.read_comsol_header(path + type)
-    assert domain_data == actual_domain_data
-
-
-@pytest.mark.parametrize("filename",  [
-    'intensity_average',
-    'intensity_only',
-    'intensity_parametric',
-    # 'level_only',
-    'pressure_acceleration_parametric_time',
-    'pressure_only',
-    'pressure_parametric',
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_data_domain(filename, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, domain_data \
-        = io.read_comsol_header(path + type)
-    data, coordinates = io.read_comsol(path + type)
-    assert domain == data.domain
-    if domain == 'freq':
-        assert all(domain_data == data.frequencies)
-    else:
-        assert all(domain_data == data.times)
-
-
-@pytest.mark.parametrize("filename,nodes",  [
-    ('intensity_average', 1),
-    ('intensity_only', 8),
-    ('intensity_parametric', 8),
-    # ('level_only', 8),
-    ('pressure_acceleration_parametric_time', 46),
-    ('pressure_only', 8),
-    ('pressure_parametric', 8),
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_data_shapes(filename, nodes, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, domain_data \
-        = io.read_comsol_header(path + type)
-    data, coordinates = io.read_comsol(path + type)
-    if coordinates is not None:
-        assert data.cshape[0] == coordinates.cshape[0]
-    assert data.cshape[0] == nodes
-    assert data.cshape[1] == len(expressions)
-    for i, para in enumerate(parameters):
-        assert data.cshape[2+i] == len(parameters[para])
-
-
-@pytest.mark.parametrize("filename",  [
-    'intensity_only',
-    'intensity_parametric',
-    'level_only',
-    'pressure_only',
-    'pressure_parametric',
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_coordinates(filename, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    data, coordinates = io.read_comsol(path + type)
-    xyz = coordinates.get_cart()
-    assert all(xyz[0] == [-.5, -.5, -.5])
-    assert all(xyz[1] == [0.5, -0.5, -0.5])
-    assert all(xyz[2] == [-0.5, 0.5, -0.5])
-    assert all(xyz[3] == [0.5, 0.5, -0.5])
-    assert all(xyz[4] == [-0.5, -0.5, 0.5])
-    assert all(xyz[5] == [0.5, -0.5, 0.5])
-    assert all(xyz[6] == [-0.5, 0.5, 0.5])
-    assert all(xyz[7] == [0.5, 0.5, 0.5])
-
-
-@pytest.mark.parametrize("filename,p1",  [
-    ('intensity_average', np.float64(1.2867253973047096E-24)),
-    ('intensity_only', np.float64(-1.0535540378988276E-8)),
-    ('intensity_parametric', np.float64(-1.0535540378988267E-8)),
-    # ('level_only', np.float64(42.754551872360125)),
-    ('pressure_acceleration_parametric_time',
-     np.float64(2.2469262686436532E-14)),
-    ('pressure_only',
-     np.complex128(complex(-3.308489057665816E-5, -0.003883799752478906))),
-    ('pressure_parametric',
-     np.complex128(complex(-3.3084890576658145E-5, -0.003883799752478905))),
-    ])
-@pytest.mark.parametrize("suffix",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_first_value_data(filename, p1, suffix):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, domain_data \
-        = io.read_comsol_header(path + suffix)
-    data, coordinates = io.read_comsol(path + suffix)
-    if domain == 'freq':
-        assert data.freq.flatten()[0] == p1
-        # assert type(data.freq.flatten()[0]) == type(p1)
-        # returns error due to a bug in FrequencyData
-    else:
-        assert data.time.flatten()[0] == p1
-        assert type(data.time.flatten()[0]) == type(p1)
-
-
-@pytest.mark.parametrize("filename,p1",  [
-    ('intensity_parametric',
-     -1.0533947432793473E-8),
-    ('pressure_parametric',
-     complex(-1.6519703918208651E-4, -0.003880099599610891)),
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_another_value_data(filename, p1, type):
-    """Test freq=500; theta=0.7854; phi=3.1416"""
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    data, coordinates = io.read_comsol(path + type)
-    point = 0
-    exp = 0
-    freq = 1
-    theta = 1
-    phi = 2
-    assert data.freq[point, exp, theta, phi, freq] == p1
-
-
-@pytest.mark.parametrize("filename,p1",  [
-    ('intensity_parametric',
-     -1.0533947432793473E-8),
-    ('pressure_parametric',
-     complex(-1.6519703918208651E-4, -0.003880099599610891)),
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_parameters_another_value_data(filename, p1, type):
-    """Test freq=500; theta=0.7854; phi=3.1416"""
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, domain_data \
-        = io.read_comsol_header(path + type)
-    theta = 1
-    phi = 2
-    parameters['theta'] = [parameters['theta'][theta]]
-    parameters['phi'] = [parameters['phi'][phi]]
-    data, coordinates = io.read_comsol(path + type, parameters=parameters)
-    point = 0
-    exp = 0
-    freq = 1
-    assert data.freq[point, exp, 0, 0, freq] == p1
-
-
-@pytest.mark.parametrize("filename",  [
-    'intensity_average',
-    'intensity_only',
-    'intensity_parametric',
-    'pressure_acceleration_parametric_time',
-    'pressure_only',
-    'pressure_parametric',
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_expressions(filename, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, domain_data \
-        = io.read_comsol_header(path + type)
-    data, coordinates = io.read_comsol(
-        path + type, expressions=[expressions[0]])
-    assert data.cshape[1] == 1
-
-
-@pytest.mark.parametrize("filename",  [
-    'pressure_parametric_incomplete',
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_check_incomplete(filename, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, domain_data \
-        = io.read_comsol_header(path + type)
-    with pytest.warns(Warning, match='inconsistent'):
-        data, coordinates = io.read_comsol(
-            path + type, expressions=[expressions[0]])
-    assert all(np.isnan(data.freq[:, 0, -1, 1]))
-    data.freq[:, 0, -1, 1] = 0
-    assert any(np.isnan(data.freq.flatten()) == False)
-
-
-@pytest.mark.parametrize("filename",  [
-    'intensity_parametric',
-    'intensity_average',
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_parameters_shape(filename, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, domain_data \
-        = io.read_comsol_header(path + type)
-    parameters['theta'] = parameters['theta'][:1]
-    parameters['phi'] = parameters['phi'][:2]
-    data, coordinates = io.read_comsol(
-        path + type, parameters=parameters)
-    assert data.cshape[1] == len(expressions)
-    assert data.cshape[2] == len(parameters['theta'][:1])
-    assert data.cshape[3] == len(parameters['phi'][:2])
-    assert data.n_bins == 2
-
-
-@pytest.mark.parametrize("filename",  [
-    'intensity_parametric',
-    'intensity_average',
-    ])
-@pytest.mark.parametrize("type",  ['.txt', '.dat', '.csv'])
-def test_read_comsol_expressions_value(filename, type):
-    path = os.path.join(os.getcwd(), 'tests', 'test_io_data', filename)
-    expressions, expressions_unit, parameters, domain, domain_data \
-        = io.read_comsol_header(path + type)
-    data_exp, coordinates = io.read_comsol(
-        path + type, expressions=[expressions[0], expressions[2]])
-    data, coordinates = io.read_comsol(path + type)
-    assert data_exp.freq[0, 0, 1, 1, 1] == data.freq[0, 0, 1, 1, 1]
-    assert data_exp.freq[0, 1, 1, 1, 1] == data.freq[0, 2, 1, 1, 1]
-    assert data_exp.freq[0, 0, 0, 2, 1] == data.freq[0, 0, 0, 2, 1]
-    assert data_exp.freq[0, 1, 0, 2, 1] == data.freq[0, 2, 0, 2, 1]
